+++ gnomad_qc/v4/resources/basics.py
@@ -1,7 +1,5 @@
 """Script containing generic resources."""
 import logging
-from collections import defaultdict
-from typing import Any, Dict, List, Optional, Union
 
 import hail as hl
 from gnomad.resources.resource_utils import (
@@ -10,7 +8,6 @@ from gnomad.resources.resource_utils import (
     VersionedTableResource,
     VersionedVariantDatasetResource,
 )
-from gnomad.utils.file_utils import check_file_exists_raise_error
 
 from gnomad_qc.v4.resources.constants import CURRENT_VERSION
 from gnomad_qc.v4.resources.meta import meta
@@ -19,7 +16,8 @@ logger = logging.getLogger("basic_resources")
 logger.setLevel(logging.INFO)
 
 
-# Note: Unlike previous versions, the v4 resource directory uses a general format of gs://gnomad/v4.0/<module>/<exomes_or_genomes>/. # noqa
+# Note: Unlike previous versions, the v4 resource directory uses a general format of
+#   gs://gnomad/v4.0/<module>/<exomes_or_genomes>/.
 def get_gnomad_v4_vds(
     split: bool = False,
     remove_hard_filtered_samples: bool = True,
@@ -314,270 +312,3 @@ def calling_intervals(
         return TableResource(
             f"gs://gnomad/resources/intervals/xgen.pad{calling_interval_padding}.dsp.pad{calling_interval_padding}.intersection.interval_list.ht"
         )
-
-
-def check_resource_existence(
-    input_step_resources: Optional[Dict[str, List]] = None,
-    output_step_resources: Optional[Dict[str, List]] = None,
-    overwrite: bool = False,
-) -> None:
-    """
-    Check the existence of all specified input and output resources.
-
-    If any of the input resources (`input_step_resources` values) don't exist, an error
-    will be raised indicating which input resources are missing.
-
-    If any of the output resources (`output_step_resources` values) already exist and
-    the `overwrite` parameter is not set to True, an error will be raised indicating
-    which output resources already exist.
-
-    If no parameters are passed to the function, nothing is done.
-
-    :param input_step_resources: A dictionary with keys as pipeline steps that generate
-        input files and the value as a list of the input files to check the existence
-        of. Default is None.
-    :param output_step_resources: A dictionary with keys as pipeline step that generate
-        output files and the value as a list of the output files to check the existence
-        of. Default is None.
-    :param overwrite: The overwrite parameter used when writing the output files.
-        Default is False.
-    :return: None.
-    """
-    # Check if the input resources exist
-    if input_step_resources:
-        for step, input_resources in input_step_resources.items():
-            check_file_exists_raise_error(
-                [r if isinstance(r, str) else r.path for r in input_resources],
-                error_if_not_exists=True,
-                error_if_not_exists_msg=(
-                    f"Not all input resources exist. Please add {step} to the command "
-                    "line. The following files are missing: "
-                ),
-            )
-
-    # Check if the output resources exist when `overwrite` is False
-    if not overwrite and output_step_resources:
-        for step, output_resources in output_step_resources.items():
-            check_file_exists_raise_error(
-                [r if isinstance(r, str) else r.path for r in output_resources],
-                error_if_exists=True,
-                error_if_exists_msg=(
-                    f"Some of the output resources that will be created by {step} "
-                    "already exist and the --overwrite argument was not set. Please "
-                    f"rerun {step} with --overwrite. The following files already exist:"
-                ),
-            )
-
-
-class PipelineStepResourceCollection:
-    """A collection of resources used in a specific step of a gnomad_qc pipeline."""
-
-    def __init__(
-        self,
-        step_name: str,
-        previous_pipeline_steps: List = [],
-        output_resources: Optional[Dict[str, Any]] = None,
-        input_resources: Optional[Dict[str, Union[Dict[str, Any], List[Any]]]] = None,
-        add_input_resources: Optional[
-            Dict[str, Union[Dict[str, Any], List[Any]]]
-        ] = None,
-        pipeline_name: Optional[str] = None,
-        overwrite: Optional[bool] = None,
-    ) -> None:
-        """
-        Construct all the necessary attributes for the object.
-
-        :param step_name: Name to be assigned to current step. The recommendation is to
-            use the step's argparse call as the name.
-        :param previous_pipeline_steps: Previous steps of the pipeline that have input
-            needed for the current step. Default is an empty List.
-        :param output_resources: Optional Dictionary of the pipeline step's output
-            resources. Keyed by the name to use as an attribute name and with the
-            associated resource as the value.
-        :param input_resources: Optional Dictionary of the pipeline step's input
-            resources. Keyed by a string used to describe where the resource comes
-            from or is created for example "generate_qc_mt.py --generate-qc-meta". The
-            Values can be either a list of resources, or a Dictionary keyed by each
-            resource's name (use as an attribute name) and with the associated resource
-            as the value. By default, the output from steps listed in
-            `previous_pipeline_steps` will be used as input resources to this step, but
-            if input_resources is supplied, then only those resources will be used.
-        :param add_input_resources: Optional Dictionary of additional input resources
-            to add for this step on top of either `input_resources` or the output from
-            steps listed in `previous_pipeline_steps`. Keyed by a string used to
-            describe where the resource comes from or is created for example:
-            "generate_qc_mt.py --generate-qc-meta". The Values can be either a list of
-            resources, or a Dictionary keyed by each resource's name (use as an
-            attribute name) and with the associated resource as the value.
-        :param pipeline_name: Optional name of the full pipeline this step belongs to.
-        :param overwrite: Whether these resources can be overwritten. Used in
-            `check_resource_existance`.
-        :return: None.
-        """
-        self.pipeline_name = pipeline_name
-        self.overwrite = overwrite
-        self.pipeline_step = step_name
-        self.previous_steps = previous_pipeline_steps
-        self._output_resource_dict = output_resources
-
-        if output_resources is not None:
-            self._add_output_resource_attributes(output_resources)
-            output_resources = {self.pipeline_step: output_resources.values()}
-
-        self.output_resources = output_resources
-
-        self.input_resources = defaultdict(list)
-        if input_resources is None:
-            for step in previous_pipeline_steps:
-                self.input_resources.update(step.output_resources)
-                self._add_output_resource_attributes(step._output_resource_dict)
-        else:
-            self.add_input_resources(input_resources)
-
-        if add_input_resources is not None:
-            self.add_input_resources(add_input_resources)
-
-    def __getattr__(self, name: str) -> None:
-        """Raise AttributeError with modified message if the attribute is not found."""
-        AttributeError(
-            f"Pipeline step {self.pipeline_step} has no resource attribute named"
-            f" {name}!"
-        )
-
-    def _add_output_resource_attributes(
-        self,
-        output_resources: Dict[str, Any],
-    ) -> None:
-        """
-        Add output resources as class attributes.
-
-        :param output_resources: Dictionary of the pipeline step's output resources.
-            Keyed by the name to use as an attribute name and with the associated
-            resource as the value.
-        :return: None.
-        """
-        for name, resource in output_resources.items():
-            setattr(self, name, resource)
-
-    def add_input_resources(
-        self,
-        input_resources: Dict[str, Union[Dict[str, Any], List[Any]]],
-    ) -> None:
-        """
-        Add input resources to the pipeline step.
-
-        :param input_resources: Dictionary of resources to add as input for this
-            pipeline step. Keyed by a string used to describe where the resource comes
-            from or is created for example: "generate_qc_mt.py --generate-qc-meta". The
-            Values can be either a list of resources, or a Dictionary keyed by each
-            resource's name (use as an attribute name) and with the associated resource
-            as the value.
-        :return: None.
-        """
-        for step, resources in input_resources.items():
-            if isinstance(resources, dict):
-                for name, resource in resources.items():
-                    setattr(self, name, resource)
-                resources = resources.values()
-            self.input_resources[step].extend(resources)
-
-    def check_resource_existance(self, overwrite: bool = None) -> None:
-        """
-        Check for existence of input and output resources.
-
-        An error is returned if input resources don't exist, or if overwrite is False
-        and an output resource does exist.
-
-        :param overwrite: Whether these resources can be overwritten.
-        :return: None.
-        """
-        if overwrite is None:
-            if self.overwrite is None:
-                overwrite = False
-            else:
-                overwrite = self.overwrite
-        check_resource_existence(
-            input_step_resources=self.input_resources,
-            output_step_resources=self.output_resources,
-            overwrite=overwrite,
-        )
-
-
-class PipelineResourceCollection:
-    """A collection of PipelineStepResourceCollection used in a gnomad_qc pipeline."""
-
-    def __init__(
-        self,
-        pipeline_name: str,
-        pipeline_steps: Dict[str, PipelineStepResourceCollection] = {},
-        pipeline_resources: Optional[Dict[str, Any]] = None,
-        overwrite: bool = None,
-    ) -> None:
-        """
-        Construct all the necessary attributes for the object.
-
-        :param pipeline_name: Name of the pipeline.
-        :param pipeline_steps: Dictionary of all steps of the pipeline Keyed by a name
-            to use for the step (added as an attribute of the object) and with the
-            associated PipelineStepResourceCollection as the Value. Default is an empty
-            Dictionary.
-        :param pipeline_resources: Optional Dictionary of resources used by multiple
-            steps of the pipeline. Keyed by the name to use as an attribute name and
-            with the associated resource as the value.
-        :param overwrite: Whether these resources can be overwritten. Used in
-            `check_resource_existance`.
-        :return: None.
-        """
-        self.pipeline_name = pipeline_name
-        self.pipeline_steps = {}
-        self.add_steps(pipeline_steps)
-        self.overwrite = overwrite
-
-        if pipeline_resources is not None:
-            for name, resource in pipeline_resources.items():
-                setattr(self, name, resource)
-
-    def __getattr__(self, name: str) -> None:
-        """Raise AttributeError with modified message if the attribute is not found."""
-        AttributeError(
-            f"Pipeline {self.pipeline_name} has no resource or pipeline step attribute"
-            f" named {name}!"
-        )
-
-    def add_steps(self, steps: Dict[str, PipelineStepResourceCollection]) -> None:
-        """
-        Add PipelineStepResourceCollections to the object.
-
-        :param steps: Dictionary of steps to add to the pipeline Keyed by a name
-            to use for the step (added as an attribute of the object) and with the
-            associated PipelineStepResourceCollection as the Value.
-        :return: None.
-        """
-        for step_name, step in steps.items():
-            step.pipeline_name = self.pipeline_name
-            step.overwrite = self.overwrite
-            setattr(self, step_name, step)
-        self.pipeline_steps.update(steps)
-
-    def check_resource_existance(
-        self, step: str, overwrite: Optional[bool] = None
-    ) -> None:
-        """
-        Check for existence of input and output resources.
-
-        An error is returned if input resources don't exist, or if overwrite is False
-        and an output resource does exist.
-
-        :param step: Name of step to check resource existence of.
-        :param overwrite: Whether these resources can be overwritten. By default, the
-            class attribute of overwrite will be used if it exists, otherwise overwrite
-            will be set to False.
-        :return: None.
-        """
-        if overwrite is None:
-            if self.overwrite is None:
-                overwrite = False
-            else:
-                overwrite = self.overwrite
-
-        self.pipeline_steps[step].check_resource_existance(overwrite=overwrite)
